+++ src/lsm/error.rs
